refactor: extract the GitHub client into internal/gh
Fourth leaf of the package split: the Client interface (was
GitHubClient), the go-github-backed RESTClient (was GoGitHubClient),
the etag caching transport, the dry-run wrapper, GitHub App auth, and
the GitHub domain types (Issue, PR, ReviewComment, PRReview, CheckRun,
WorkflowRun, WorkflowJob) move to internal/gh with their tests.
Agent-side types (IssueWork, AgentResult, Commit, JobRun) stay in
pkg/agent, which bridges all old names with aliases and wrapper funcs.

Two dependencies the client had quietly grown on agent internals are
severed: the CountCommitsSince pagination short-circuit gets its own
documented cap instead of reaching for the agent's rebase threshold,
and the dry-run logger uses a local rune-safe truncate helper.
